Add caching decorator from drf-extensions
This does not address the environment-based caching control
that we had talked about, but it's so easy to implement that
it seemed worth testing and getting eyes on.

Add REST_FRAMEWORK_EXTENSIONS settings to settings.py

Specifically, add settings specific to the caching behavior.
For now, I think we can keep these settings at their default values,
so the only thing I changed was to disable caching for errors.

Add custom cache key construction

Added custom key construction to ensure that POST request
params are taken into account when hitting the cache

Add distinct Django cache for DRF-caching

Making a separate cache for DRF-extensions will give us
more flexibility for disabling the API cache as needed
(during development, for example)

Override viewset retrieve method to apply caching decorator

The front-end team noticed that performance on individual
awards was still very poor. It turns out that those urls
route to the viewset retrieve method (which makes sense). Our
code doesn't override the retireve method, thus we weren't
applying the caching decorator to. This fixes that. There's
probably a more elegant way to do this, but let's see if
this helps for now.